Added gaussian, clopper_pearson files and removed type from cb interface
